GenerateCNNTraining.C
This C based program creates a trivial spectrum of gamma rays with background features such as compton and pair production. This was created to build a training dataset for UNET ML for Noisy spectrum cleaning. This is only a simple minded solution to training our neural network to determine what noise is within our datasets.